The problem was in module get-folder-size. For some f*in reason it was interrupting the whole app from launching. So, just took main functionality from it and put into helper. New checkpoint, app launches. !Also changed storage of free space in db from gb to mb.

+++ src/database/models/group-path.model.ts
@@ -9,10 +9,10 @@ export class GroupPath implements IGroupPath {
 	@Column("text")
 	pathName!: string;
 
-	@Column("double")
+	@Column("float8")
 	sizeMax!: number;
 
-	@Column("double")
+	@Column("float8")
 	sizeUsed!: number;
 
 	@Column("boolean")
